rename StaticMethodReplacerRector to RenameStaticMethodRector
