+++ config/level/cakephp/cakephp36.yaml
@@ -1,7 +1,7 @@
 # source: https://book.cakephp.org/3.next/en/appendices/3-6-migration-guide.html
 
 services:
-    Rector\Rector\MethodCall\MethodNameReplacerRector:
+    Rector\Rector\MethodCall\RenameMethodRector:
         'Cake\ORM\Table':
             'association': 'getAssociation'
         'Cake\Validation\ValidationSet':
