+++ config/level/sylius/sylius10.yaml
@@ -3,7 +3,7 @@
 services:
     Rector\Sylius\Rector\Review\ReplaceCreateMethodWithoutReviewerRector: ~
 
-    Rector\Rector\MethodCall\MethodNameReplacerRector:
+    Rector\Rector\MethodCall\RenameMethodRector:
         'Sylius\Component\Core\Repository\OrderRepositoryInterface':
             'count': 'countPlacedOrders'
             'countByChannel': 'countFulfilledByChannel'
