+++ config/level/symfony/symfony41.yaml
@@ -1,6 +1,6 @@
 services:
     # https://github.com/symfony/symfony/blob/master/UPGRADE-4.1.md
-    Rector\Rector\MethodCall\MethodNameReplacerRector:
+    Rector\Rector\MethodCall\RenameMethodRector:
         # https://github.com/symfony/symfony/commit/463f986c28a497571967e37c1314e9911f1ef6ba
         Symfony\Component\Console\Helper\TableStyle:
             setHorizontalBorderChar: 'setHorizontalBorderChars'
