+++ tests/Rector/MethodCall/RenameStaticMethodRector/Fixture/fixture.php.inc
