+++ tests/Rector/MethodCall/RenameStaticMethodRector/Fixture/fixture2.php.inc
