+++ tests/Rector/MethodCall/RenameStaticMethodRector/Source/FormMacros.php
